FIX: Repair Day7 changes that were overwritten

+++ src/common/mod.rs
@@ -2,12 +2,12 @@ use std::fs;
 use std::fs::File;
 use std::io::{BufRead, BufReader};
 
-pub fn import(name: &str) -> Vec<i32> {
+pub fn import(name: &str) -> Vec<i64> {
 	let file = File::open(name).unwrap();
 	let reader = BufReader::new(file);
 	let mut vec = Vec::new();
 	for (_, line) in reader.lines().enumerate() {
-		vec.push(line.unwrap().parse::<i32>().unwrap())
+		vec.push(line.unwrap().parse::<i64>().unwrap())
 	}
 	return vec;
 }
